proxier/ipvs: Retrieve all IPVS services in a single call
This commit allows kube-proxy to retrieve all IPVS services in a single
netlink call to the kernel. Previously, at least two `ipvs.GetService`
calls would be issued for each ClusterIP service, more for other types
of service. In environments with a lot of services this would mean a
huge number of context switches.

This commit leverages the fact that it's possible to get all ipvs
services from the system in a single `ipvs.GetServices` call. This
function was even already used in the `syncProxyRules` code to cleanup
old services, this commit simply moves it to an earlier point and
ensures we keep a map of cached services updated with the ones the sync
process needs to create.

Signed-off-by: Cezar Sa Espinola <cezarsa@gmail.com>